requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.20
alembic==1.11.1
pydantic==1.10.12
paho-mqtt==1.6.1
python-dotenv==1.0.0
joblib==1.3.2
numpy==1.26.4
pandas==2.2.2